+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added cluster support for `XADD`, `XDEL` and `XRANGE` (#1587)
 - Added prefixable interface for `HEXPIRE` and `HEXPIRETIME` (#1592)
 - Added temporary XREADGROUP_CLAIM command (#1608)
+- Added support for MSET command (#1610)
 
 ### Changed
 - Refactor pipeline data writing depends on connection type (#1586)
@@ -19,6 +20,7 @@
 ### Added
 - Added support for `XDELEX` and `XACKDEL` (#1580)
 - Added missing VSIM argument (#1582)
+- Added experimental support for FT.HYBRID (#1607)
 
 ### Changed
 - Extended `XTRIM` and `XADD` commands with new parameters (#1580)